# Streaming PromQL engine: add support for range vector selectors as the top-level expression