--- a/pkg/streamingpromql/query.go
+++ b/pkg/streamingpromql/query.go
@@ -11,6 +11,7 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/model/timestamp"
 	"github.com/prometheus/prometheus/promql"
 	"github.com/prometheus/prometheus/promql/parser"
@@ -351,6 +352,10 @@ func (q *Query) populateMatrixFromRangeVectorOperator(ctx context.Context, o ope
 			Floats: b.CopyPoints(),
 		})
 	}
+
+	slices.SortFunc(m, func(a, b promql.Series) int {
+		return labels.Compare(a.Metric, b.Metric)
+	})
 
 	return m, nil
 }