… CHANGELOG.md
… pkg/streamingpromql/README.md
… pkg/streamingpromql/benchmarks/benchmarks.go
… pkg/streamingpromql/engine_test.go
… pkg/streamingpromql/operator/aggregation_test.go
… pkg/streamingpromql/operator/range_vector_selector_with_transformation.go
… pkg/streamingpromql/operator/ring_buffer_test.go
… pkg/streamingpromql/operator/selector.go
… pkg/streamingpromql/testdata/ours/functions.test